Adds day by day header to the incidents view
This changes how the display of incidents are rendered by adding a day
by day date header. On days without any incidents, the date header is
shown as well as a "No reported incidents" text. On days with
incidents, the view will add the date header in front of the first
incident on the given day.

- New function render_no_incident added to incident class which renders
  days with no registered incidents.